+++ tests/test_benchmarks.py
@@ -31,7 +31,7 @@ def test_token_parsing(benchmark):
         for i in range(10000):
             s += 'uint:12=244, float:32=0.4'
             s += '0x3e44f, 0b11011, 0o75523'
-            s += [0, 1, 2, 0, 0, 1, 2, 0, -1, 0, 'hello']
+            s += bitstring.BitArray.from_bools([0, 1, 2, 0, 0, 1, 2, 0, -1, 0, 'hello'])
             s += bitstring.BitArray.from_zeros(104)
     benchmark(token_parsing)